Lanjutan delete

+++ app/Http/Controllers/ThesisSeminarController.php
@@ -12,7 +12,7 @@ class ThesisSeminarController extends Controller
         $thesisseminars = DB::table('thesis_seminars')
                           ->join('theses', 'thesis_seminars.thesis_id', '=', 'theses.thesis_id')
                           ->join('thesis_proposals', 'theses.thesis_id', '=', 'thesis_proposals.thesis_id')
-                          ->select('thesis_seminars.id', 'thesis_seminars.registered_at', 'thesis_seminars.seminar_at', 'thesis_seminars.status', DB::raw('(CASE WHEN thesis_seminars.status = 1 THEN '. "'Mengajukan'".'END) AS status_semhas'))
+                          ->select('thesis_seminars.id', 'thesis_seminars.registered_at', 'thesis_seminars.seminar_at', 'thesis_seminars.status', DB::raw('(CASE WHEN thesis_seminars.status = 1 THEN '. "'Mengajukan'".' WHEN thesis_seminars.status = 2 THEN '."'Disetujui'".' END) AS status_semhas'))
                           ->paginate(3);
 
         return view('backend.thesis_seminar.index', compact('thesisseminars'));
@@ -20,18 +20,18 @@ class ThesisSeminarController extends Controller
     
     public function create()
     {
-      /*$ruangan = Ruangan::all()->pluck('nama','id');
-      $mahasiswa = DB::table('ta_semhas')
-                  ->join('ta_sempro', 'ta_semhas.ta_sempro_id', '=' , 'ta_sempro.id')
-                  ->join('tugas_akhir','ta_sempro.tugas_akhir_id', '=', 'tugas_akhir.id')
-                  ->join('mahasiswa', 'tugas_akhir.mahasiswa_id', '=', 'mahasiswa.id')
-                  ->select('ta_sempro.id','mahasiswa.nama')
-                  ->pluck('mahasiswa.nama','ta_semhas.id');
-      $rekomendasi = DB::table('ta_semhas')
+      $student = DB::table('thesis_seminars')
+                ->join('theses', 'thesis_seminars.thesis_id', '=', 'theses.thesis_id')
+                ->join('thesis_proposals', 'theses.thesis_id', '=', 'thesis_proposals.thesis_id')
+                ->join('students', 'theses.student_id', '=', 'students.id')
+                ->select('theses.id')
+                ->get();
+
+      /*$rekomendasi = DB::table('ta_semhas')
                     ->select('rekomendasi', DB::raw('(CASE WHEN rekomendasi = 1 THEN '. "'Mengulang Seminar'" .'WHEN rekomendasi = 2 THEN '. "'Lanjut Sidang dengan Revisi'".'WHEN rekomendasi = 3 THEN '."'Lanjut Sidang Tanpa Revisi'".'END) AS rekomendasi_semhas'))
                     ->distinct()
                     ->pluck('rekomendasi_semhas','rekomendasi');*/
-      return view('backend.thesis_seminar.create');
+      return view('backend.thesis_seminar.create', compact('student'));
     }
 
     public function store(Request $request)
@@ -45,12 +45,12 @@ class ThesisSeminarController extends Controller
             $semhas->thesis_id = $request->thesis_id;
             $semhas->status = $request->status;
       
-            if($request->hasFile('file_laporan') && $request->file('file_report')->isValid())
+            if($request->hasFile('file_report') && $request->file('file_report')->isValid())
             {
                 $filename = uniqid('laporan-');
                 $fileext = $request->file('file_report')->extension();
                 $filenameext = $filename.'.'.$fileext;
-                $filepath = $request->file_report->storeAs('/laporan_ta',$filenameext);
+                $filepath = $request->file_report->storeAs('public/laporan_ta',$filenameext);
                 $semhas->file_report = $filepath;
             }
             if($semhas->save()) {
@@ -82,18 +82,32 @@ class ThesisSeminarController extends Controller
 
     public function destroy($id)
     {
-        $a = db::table('thesis_sem_audiences')
-        ->where('thesis_seminar_id','=',$id);
-        $a->delete();
-        $b = db::table('thesis_sem_reviewers')
-        ->where('thesis_seminar_id','=',$id);
-        $b->delete();
-        $thesisseminars = DB::table('thesis_seminars')
-        // ->join('thesis_sem_audiences', 'thesis_seminars.id', '=', 'thesis_sem_audiences.thesis_seminar_id')
-        // ->join('thesis_sem_reviewers', 'thesis_seminars.id', '=', 'thesis_sem_reviewers.thesis_seminar_id')
-        ->where('thesis_seminars.id','=',$id);
-        $thesisseminars->delete();
-        session()->flash('flash_success', 'Berhasil membatalkan pengajuan semhas');
-        return redirect()->route('admin.semhas.index');
+        $statuss = DB::table('thesis_seminars') 
+                  ->select('thesis_seminars.status')
+                  ->where('thesis_seminars.id','=', $id)
+                  ->get();
+        //echo $status;
+        foreach($statuss as $status)
+        {
+            foreach($status as $st)
+            {
+                if($st == 1)
+                {
+                    $a = DB::table('thesis_sem_audiences')->where('thesis_seminar_id','=',$id);
+                    $a->delete();
+                    $b = DB::table('thesis_sem_reviewers')->where('thesis_seminar_id','=',$id);
+                    $b->delete();
+                    $thesisseminars = DB::table('thesis_seminars')->where('thesis_seminars.id','=',$id);
+                    $thesisseminars->delete();
+                    session()->flash('flash_success', 'Berhasil membatalkan pengajuan semhas');
+                    return redirect()->route('admin.semhas.index');
+            }
+                elseif($st != 1)
+                {
+                    session()->flash('flash_success', 'Gagal membatalkan pengajuan. Pengajuan telah disetujui.');
+                    return redirect()->route('admin.semhas.index');
+                }
+            }
+        }
     }
 }
